dos.library: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39326 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/dos/internalloadseg_aos.c
@@ -492,7 +492,7 @@ done:
         /* overlay executable */
         h[3] = (ULONG)fh;
         h[4] = (ULONG)overlaytable;
-        h[5] = MKBADDR(hunktab);
+        h[5] = (ULONG)MKBADDR(hunktab);
         D(bug("overlay loaded!\n"));
         return (BPTR)(-(LONG)MKBADDR(h));
 #endif
